Optimisations regarding LinkedList and the way game objects are stored in World and GameWorld, plus a few other things (improvements for Profiler and Application, specifically).

+++ src/com/stabilise/util/maths/MathsUtil.java
@@ -377,7 +377,7 @@ public class MathsUtil {
 	 * @return The compacted value.
 	 */
 	public static long compactLong(int x, int y) {
-		return (long)((x << 32) & y);
+		return ((long)x << 32) & (long)y;
 	}
 	
 }
